changed as requested in pr #135
