+++ .vscode/launch.json
@@ -8,7 +8,7 @@
         "type": "chrome",
         "request": "launch",
         "name": "vuejs: chrome",
-        "url": "http://localhost:8091",
+        "url": "http://localhost:8093",
         "webRoot": "${workspaceFolder}/src",
         "breakOnLoad": true,
         "sourceMapPathOverrides": {
