+++ public/index.html
@@ -99,11 +99,6 @@
   <div id="app"></div>
 
   <div id="loading">
-    <div class="spinner">
-      <!-- <div class="bounce1"></div>
-      <div class="bounce2"></div>
-      <div class="bounce3"></div> -->
-    </div>
     <img src="/img/yesodot_logo.svg" alt="" class="loading-logo">
   </div>
 </body>
