+++ src/api/index.js
@@ -4,6 +4,8 @@ import * as users from './users'
 import * as delegators from './delegators'
 import * as settings from './settings'
 import * as quota from './quota'
+import * as exShare from './exShare'
+import * as utils from './utils'
 import search from './search'
 import commands from './commands'
 
@@ -12,8 +14,10 @@ export {
   share,
   users,
   delegators,
+  exShare,
   settings,
   commands,
   search,
-  quota
+  quota,
+  utils
 }
