+++ src/api/users.js
@@ -71,31 +71,4 @@ export async function remove (id) {
   }
 }
 
-export function simplifyStatus(status) {
-	switch (status) {
-    case "REQUEST_NOT_APPROVED":
-      return "denied";
-    case "REQUEST_PENDING_APPROVAL":
-      return "waiting for approval";
-    case "FILE_ABORTED":
-      return 'failed'
-    case "FILE_DROPPED":
-      return 'failed'
-    case "FILE_TRANSFER_FAILED":
-      return 'failed'
-    case "FILE_INVALID_HASH":
-      return 'failed';
-    case "FILE_DELETED":
-      return 'failed';
-    case "FILE_PASSED":
-      return 'passed';
-    case "FILE_TRANSFER_TO_OUT_FOLDER":
-      return 'passed';
-    case "FILE_FIXED":
-      return 'passed';
-    case "FILE_FAILED_BAD_SUFFIX":
-      return "bad suffix";
-    default:
-      return 'sending';
-	}
-}
+
